Fix error

+++ background_threads.py
@@ -61,14 +61,13 @@ def track_ip():
 # thread checks first if the shard is in active contracts, if no it deletes the shard
 def withdraw():
     shards = os.listdir(settings.data_directory)
-    active_shards = api_requests.get_active_contracts()
-    active_shards = active_shards["shards"]
-    print(active_shards)
+    active_shards, success = api_requests.get_active_contracts()
 
     for shard in shards:
-        if shard not in active_shards:
-            os.unlink(os.path.join(settings.data_directory, shard))
-            continue
+        if success:
+            if shard not in active_shards:
+                os.unlink(os.path.join(settings.data_directory, shard))
+                continue
 
         api_requests.withdraw_request(shard)
